fix a bug in tune_params, add train-test regression tests
they work but for some reason are waaaaay slow...need to look into that

+++ sdm/tune_params.hpp
@@ -338,7 +338,7 @@ std::pair<size_t, size_t> tune_params(
             boost::rethrow_exception(errors[i]);
 
     // get all the best configs into one vector
-    size_t best_score = *std::max_element(
+    double best_score = *std::max_element(
             scores.begin(), scores.end());
     std::vector<config> best_configs;
